Refactor the encode function into smaller functions

+++ mg_bitstream.h
@@ -25,8 +25,8 @@ struct bitstream {
 };
 
 void Rewind(bitstream* Bs);
-size_t Size(const bitstream& Bs); // TODO: takes a non-pointer as argument
-size_t BitSize(const bitstream& Bs);
+i64 Size(const bitstream& Bs); // TODO: takes a non-pointer as argument
+i64 BitSize(const bitstream& Bs);
 /* ---------------- Read functions ---------------- */
 void InitRead(bitstream* Bs, buffer Stream);
 /* Refill our buffer (replace the consumed bytes with new bytes from memory) */
